DatasetManager mixin refactored into other components (indicatorsModel and viewConfReader). Issue with preloaded data of localidadeview solved. Run some tests before merging into master.
